requirements.txt
requests==2.27.1
schedule==1.1.0
python-telegram-bot==13.12
SQLAlchemy==1.4.37
